refactor: extract update alert dialog into component
